Create KosarajusAlgorithm.java
I have added the Kosaraju's Algorithm. Kosaraju's Algorithm is an efficient method to find all strongly connected components (SCCs) in a directed graph. A strongly connected component is a maximal subgraph where any two vertices are reachable from each other. This algorithm works in O(V + E) time complexity, where V is the number of vertices and E is the number of edges.