Create primsAlgorithm.java
I have added the Prims Algorithm. Prim's Algorithm is a greedy algorithm used to find the Minimum Spanning Tree (MST) of a connected, undirected graph. The MST is a subset of the edges that connects all the vertices with the minimum possible total edge weight, ensuring that there are no cycles. The algorithm works by building this tree step by step, starting from an arbitrary vertex and expanding it by adding the minimum weight edge that connects a new vertex to the tree.